Fixed a horrible MegaBug and added HTTP message sending + D2D sending (no Rx yet)

+++ user/uart_buffer.h
@@ -14,7 +14,7 @@ void UART_AllocBuffers(void);
 uint16 UART_ReadAsync(char *pdata, uint16 data_len);
 
 // write to tx buffer
-void UART_SendAsync(const char *pdata, int16_t data_len);
+void UART_SendAsync(const char *pdata, int data_len);
 
 //move data from uart fifo to rx buffer
 void UART_RxFifoCollect(void);
